Database working with counter/line length
