+++ app/src/main/java/com/example/bouncermodule/MainActivity.java
@@ -6,6 +6,7 @@ import android.graphics.Bitmap;
 import android.graphics.Color;
 import android.os.Bundle;
 import android.provider.MediaStore;
+import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 import android.widget.ImageView;
@@ -13,6 +14,7 @@ import android.widget.TextView;
 
 import com.google.android.material.bottomnavigation.BottomNavigationView;
 
+import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.app.ActivityCompat;
@@ -23,6 +25,14 @@ import androidx.navigation.ui.AppBarConfiguration;
 import androidx.navigation.ui.NavigationUI;
 
 import com.example.bouncermodule.databinding.ActivityMainBinding;
+import com.google.firebase.database.DataSnapshot;
+import com.google.firebase.database.DatabaseError;
+import com.google.firebase.database.DatabaseReference;
+import com.google.firebase.database.FirebaseDatabase;
+import com.google.firebase.database.ValueEventListener;
+
+import java.util.HashMap;
+import java.util.Map;
 
 public class MainActivity extends AppCompatActivity implements View.OnClickListener {
 
@@ -42,10 +52,6 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
 
     private ImageView photoIdImageView;
     private Button photoIdButton;
-
-    // Firebase variables
-    private DatabaseReference mDatabase;
-    private Map<String, Bars> barsMap = new HashMap<>();
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -70,16 +76,6 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
 //        minusButton = (Button) findViewById(R.id.minus);
 //        minusButton.setOnClickListener(this);
 
-        currentLength = (TextView) findViewById(R.id.CurrentLengthValue);
-        counterValue = (TextView) findViewById(R.id.Total_Value);
-        plusButton = (Button) findViewById(R.id.plus);
-        plusButton.setOnClickListener(this);
-        minusButton = (Button) findViewById(R.id.minus);
-        minusButton.setOnClickListener(this);
-
-        photoIdImageView = findViewById(R.id.imageviewphotoid);
-        photoIdButton = findViewById(R.id.capturePhotoId);
-
         //Request for camera runtime permission
 //        if (ContextCompat.checkSelfPermission( MainActivity.this, Manifest.permission.CAMERA) != PackageManager.PERMISSION_GRANTED) {
 //            ActivityCompat.requestPermissions(MainActivity.this, new String[]{
@@ -89,46 +85,24 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
         // Passing each menu ID as a set of Ids because each
         // menu should be considered as top level destinations.
         AppBarConfiguration appBarConfiguration = new AppBarConfiguration.Builder(
-                R.id.navigation_home, R.id.navigation_facerecognition)
+                R.id.navigation_home, R.id.navigation_map, R.id.navigation_bars)
                 .build();
         NavController navController = Navigation.findNavController(this, R.id.nav_host_fragment_activity_main);
         NavigationUI.setupActionBarWithNavController(this, navController, appBarConfiguration);
         NavigationUI.setupWithNavController(binding.navView, navController);
 
-        mDatabase = FirebaseDatabase.getInstance().getReference();
-
-        DatabaseReference barRef = mDatabase.child("bars/");
-        barRef.addValueEventListener(new ValueEventListener(){
-            @Override
-            public void onDataChange(DataSnapshot dataSnapshot){
-                for(DataSnapshot data: dataSnapshot.getChildren()){
-                    String barName = data.getKey();
-                    String lineLength = data.child("lineLength").getValue().toString();
-                    Integer lineCount = Integer.parseInt(data.child("lineCount").getValue().toString());
-                    Double latitude = Double.parseDouble(data.child("latitude").getValue().toString());
-                    Double longitude = Double.parseDouble(data.child("longitude").getValue().toString());;
-                    Bars tempBar = new Bars(lineLength, lineCount, longitude, latitude);
-                    barsMap.put(barName, tempBar);
-                }
-            }
-
-            @Override
-            public void onCancelled(@NonNull DatabaseError error) {
-
-            }
-        });
 
     }
 
     // used for reading in file filled with new bars
-    public void writeNewBar(String barId, String lineLength, Integer lineCount, Double longitude, Double latitude) {
-        if (barsMap.containsKey(barId) == false){
-            Bars bar = new Bars(lineLength, lineCount, longitude, latitude);
-            Log.d("Writing new Bar", barId);
-            barsMap.put(barId, bar);
-            mDatabase.child("bars").child(barId).setValue(bar);
-        }
-    }
+//    public void writeNewBar(String barId, String lineLength, Integer lineCount, Double longitude, Double latitude) {
+//        if (barsMap.containsKey(barId) == false){
+//            Bars bar = new Bars(lineLength, lineCount, longitude, latitude);
+//            Log.d("Writing new Bar", barId);
+//            barsMap.put(barId, bar);
+//            mDatabase.child("bars").child(barId).setValue(bar);
+//        }
+//    }
 
     @Override
     public void onClick(View view) {
@@ -152,16 +126,7 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
                 counterValInt--;
             }
             counterValue.setText("Total:    " + String.valueOf(counterValInt));
-        }else if(view.getId() == R.id.imageviewphotoid){
-            Intent intent = new Intent(MediaStore.ACTION_IMAGE_CAPTURE);
-            startActivityForResult(intent, 100);
         }
-
-        // Changing the line count and line length of Mondays
-        Bars tempBar = barsMap.get("Mondays");
-        tempBar.setLineCount(counterValInt);
-        tempBar.setLineLength((String) currentLength.getText());
-        mDatabase.child("bars").child("Mondays").setValue(tempBar);
     }
     @Override
     protected void onActivityResult(int requestCode, int resultCode, @Nullable Intent data) {
