+++ build.gradle
@@ -1,7 +1,35 @@
-// Top-level build file where you can add configuration options common to all sub-projects/modules.
-plugins {
-    id 'com.android.application' version '7.4.0' apply false
-    id 'com.android.library' version '7.4.0' apply false
-    id 'com.google.android.libraries.mapsplatform.secrets-gradle-plugin' version '2.0.1' apply false
+ //Top-level build file where you can add configuration options common to all sub-projects/modules.
 
+// Top-level build file where you can add configuration options common to all sub-projects/modules.
+buildscript {
+    repositories {
+        // Make sure that you have the following two repositories
+        google()  // Google's Maven repository
+
+        mavenCentral()  // Maven Central repository
+
+    }
+    dependencies {
+        // Add the dependency for the Google services Gradle plugin
+        classpath 'com.android.tools.build:gradle:7.4.2'
+        classpath 'com.google.gms:google-services:4.3.15'
+
+    }
 }
+
+ plugins {
+     id 'com.android.application' version '7.4.0' apply false
+     id 'com.android.library' version '7.4.0' apply false
+     id 'com.google.android.libraries.mapsplatform.secrets-gradle-plugin' version '2.0.1' apply false
+
+ }
+
+ allprojects {
+     repositories {
+         // Make sure that you have the following two repositories
+         google()  // Google's Maven repository
+
+         mavenCentral()  // Maven Central repository
+
+     }
+ }
